# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Common Ground API
  description: |
    API for a neighbourhood community platform connecting people through clubs, events, and meals.
    Built as part of **COM5409 — Web Design & Programming**.
  version: 1.0.0
servers:
  - url: http://localhost:3005/api
    description: Local development server
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - fullName
              properties:
                email:
                  type: string
                  example: "user@example.com"
                password:
                  type: string
                  example: "password123"
                fullName:
                  type: string
                  example: "John Doe"
                ageGroup:
                  type: string
                  example: "Adult (40-60)"
                interests:
                  type: array
                  items:
                    type: string
                  example: ["Books", "Walking", "Gardening"]
      responses:
        "200":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "400":
          description: Missing fields or email exists
  /auth/signin:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: "user@example.com"
                password:
                  type: string
                  example: "password123"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials

  # ============ CLUBS ============
...
  /events:
    get:
      tags: [Events]
      summary: Get all events
      description: Returns all 20 community events
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

    post:
      tags: [Events]
      summary: Create a new event
      description: Add a new event to the community calendar
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - club
                - title
                - date
                - location
              properties:
                club:
                  type: string
                  example: "Book Club"
                title:
                  type: string
                  example: "Monthly Book Meeting"
                date:
                  type: string
                  example: "Thursday, 7pm"
                location:
                  type: string
                  example: "Community Centre"
                description:
                  type: string
                  example: "Discuss this month's book"
      responses:
        "200":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
  /events/{id}/attend:
    post:
      tags: [Events]
      summary: Attend an event
      description: Mark the authenticated user as attending an event
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Event ID
      responses:
        "200":
          description: Attended successfully
        "401":
          description: Unauthorized

  # ============ MEALS ============
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: "user@example.com"
        fullName:
          type: string
          example: "John Doe"
        phone:
          type: string
          example: "07700 900123"
        ageGroup:
          type: string
          example: "Adult (40-60)"
        interests:
          type: array
          items:
            type: string
          example: ["Books", "Walking", "Gardening"]
        joinedEvents:
          type: array
          items:
            type: integer
        requestedMeals:
          type: array
          items:
            type: integer
        joinedClubs:
          type: array
          items:
            type: integer
...
    Event:
      type: object
      properties:
        id:
          type: integer
          example: 1
        club:
          type: string
          example: "History Club"
        title:
          type: string
          example: "Local Heritage Walk"
        date:
          type: string
          example: "Saturday, 10am"
        location:
          type: string
          example: "Old Library"
        attendees:
          type: integer
          example: 14
        description:
          type: string
          example: "Guided walk through neighbourhood history."
...